pyproject.toml
[project]
name = "gemini-translate"
version = "0.1.0"
description = "CLI tool to translate text using Google Gemini API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
]

[dependency-groups]
dev = [
    "ruff>=0.12.1",
]

[project.scripts]
gtr = "gemini_translate.cli:main"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py313"

include = ["*.py", "*.pyi", "**/pyproject.toml"]
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "dist",
    "build",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
